IntelliiJ automatic code refactor and cleanup

+++ src/main/java/com/example/demo/controller/AdminController.java
@@ -1,8 +1,8 @@
 package com.example.demo.controller;
 
 import com.example.demo.model.User;
-import com.example.demo.repository.UserRepository;
 import com.example.demo.repository.ReviewRepository;
+import com.example.demo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
@@ -26,16 +26,16 @@ public class AdminController {
         List<User> users = userRepository.findAll();
         long userCount = users.size();
         long reviewCount = reviewRepository.count();
-        
+
         long activeUsers = userCount;
 
         var usersWithReviewCounts = users.stream()
-            .map(user -> {
-                var reviewsCount = reviewRepository.findByUserID(user.getId()).size();
-                user.setReviewCount((int) reviewsCount);
-                return user;
-            })
-            .collect(Collectors.toList());
+                .map(user -> {
+                    var reviewsCount = reviewRepository.findByUserID(user.getId()).size();
+                    user.setReviewCount(reviewsCount);
+                    return user;
+                })
+                .collect(Collectors.toList());
 
         model.addAttribute("users", usersWithReviewCounts);
         model.addAttribute("userCount", userCount);
